+++ tests/library/Stubs/WithUninitialized.php
@@ -1,0 +1,17 @@
+<?php
+
+/*
+ * Copyright (c) Alexandre Gomes Gaigalas <alganet@gmail.com>
+ * SPDX-License-Identifier: MIT
+ */
+
+declare(strict_types=1);
+
+namespace Respect\Validation\Test\Stubs;
+
+final class WithUninitialized
+{
+    public string $initialized = 'foo';
+
+    public string $uninitialized;
+}